Modify includes.xml for local cdart
changed the paths for the includes for cd art in the widget,list,and
now playing so it points to a cdart folder in /home/xbmc/.xbmc/userdata/cdart
that way i dont have to have cdart with the media as mine is on network and
results in a lag in loading. Noid you should cosider pulling this.